stamps_web/requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.12.0
freezegun==1.2.2
psycopg==3.1.12
psycopg-binary==3.1.12
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
wheel==0.41.2
